Factor out Patterns into separate file
