+++ src/components/TaxPayer.tsx
@@ -2,7 +2,8 @@ import React, { ReactElement } from 'react'
 import { useForm } from 'react-hook-form'
 import { Box } from '@material-ui/core'
 import { useDispatch, useSelector } from 'react-redux'
-import { LabeledInput, Patterns } from './labeledInput'
+import { LabeledInput } from './labeledInput'
+import { Patterns } from './Patterns'
 import { saveTaxpayerInfo } from '../redux/actions'
 import { PagedFormProps } from './pager'
 import { TaxesState, TaxPayer } from '../redux/data'
