+++ src/components/labeledInput.tsx
@@ -3,18 +3,13 @@ import { TextField, Box, Checkbox, FormControlLabel } from '@material-ui/core'
 import { Control, Controller, RegisterOptions } from 'react-hook-form'
 import InputMask from 'react-input-mask'
 import locationPostalCodes from '../data/locationPostalCodes'
+import { PatternConfig } from './Patterns'
 
 interface Errors {
   [key: string]: {type: string}
 }
 
 type Register = (rules?: RegisterOptions) => any
-
-interface PatternConfig {
-  mask?: string
-  regexp?: RegExp
-  description?: string
-}
 
 interface LabeledInputProps {
   strongLabel?: string
@@ -226,46 +221,4 @@ export function LabeledCheckBox (props: LabeledCheckBoxProps): ReactElement {
       control={control}
     />
   )
-}
-
-export const Patterns: {[name: string]: PatternConfig} = {
-  name: {
-    regexp: /^[A-Za-z ]+$/i,
-    description: 'Input should only include letters and spaces'
-  },
-  zip: {
-    regexp: /[0-9]{5}-[0-9]{4}/,
-    description: 'Input should be filled with 9 numbers',
-    mask: '99999-9999'
-  },
-  ssn: {
-    mask: '999-99-9999',
-    regexp: /[0-9]{3}-[0-9]{2}-[0-9]{4}/,
-    description: 'Input should be filled with 9 numbers'
-  },
-  ein: {
-    mask: '99-9999999',
-    regexp: /[0-9]{2}-[0-9]{7}/,
-    description: 'Input should be filled with 9 numbers'
-  },
-  currency: {
-    mask: '$999999',
-    regexp: /[0-9]*/,
-    description: 'Input should be filled with numbers only'
-  },
-  bankAccount: {
-    mask: '999999999999',
-    regexp: /[0-9]{10}|[0-9]{11}|[0-9]{12}/,
-    description: 'Input should be filled with 10-12 numbers'
-  },
-  bankRouting: {
-    mask: '999999999',
-    regexp: /[0-9]{9}/,
-    description: 'Input should be filled with 10 numbers'
-  },
-  usPhoneNumber: {
-    regexp: /[0-9]{3}-[0-9]{3}-[0-9]{4}/,
-    description: 'Input should be filled with 10 numbers',
-    mask: '999-999-9999'
-  }
 }
